refactor: 기존 Admin API -> AdminController로 이동
